Fixing file mess
