--- app/src/main/java/mx/itesm/bamx/activities/MainActivity.kt
@@ -1,103 +1,0 @@
-package mx.itesm.bamx
-
-import android.content.Context
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import android.view.View
-import android.widget.Button
-import android.widget.Toast
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.common.api.ApiException
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.GoogleAuthProvider
-import java.lang.Exception
-
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var firebaseAuth: FirebaseAuth
-    private lateinit var gsc : GoogleSignInClient
-    private lateinit var gso : GoogleSignInOptions
-    private lateinit var googleButton : Button
-
-    private companion object{
-        private const val RC_SIGN_IN = 100
-        private const val TAG = "GOOGLE_SIGN_IN_TAG"
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        googleButton = findViewById(R.id.googleLoginButton)
-
-        //Configure google signIn
-        gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-            .requestIdToken(getString(R.string.default_web_client_id))
-            .requestEmail()
-            .build()
-
-        gsc = GoogleSignIn.getClient(this, gso)
-
-        //Initialize Firebase auth
-        firebaseAuth = FirebaseAuth.getInstance()
-        checkUser()
-
-        googleButton.setOnClickListener{
-            val intent = gsc.signInIntent
-            startActivityForResult(intent, RC_SIGN_IN)
-        }
-
-
-    }
-
-    private fun checkUser(){
-
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == RC_SIGN_IN){
-            Log.d(TAG,"onActivityResult: GoogleSignIn intent result")
-            val accountTask = GoogleSignIn.getSignedInAccountFromIntent(data)
-            try {
-                val account = accountTask.getResult(ApiException::class.java)
-                firebaseAuthWithGoogleAccount(account)
-            } catch (e: Exception){
-                Log.d(TAG, "onActivityResult: ${e.message}")
-            }
-        }
-    }
-
-    private fun firebaseAuthWithGoogleAccount(account: GoogleSignInAccount?) {
-        Log.d(TAG, "firebaseAuthWithGoogleAccount : begin firebase auth with google account")
-        val credential = GoogleAuthProvider.getCredential(account!!.idToken, null)
-        firebaseAuth.signInWithCredential(credential)
-            .addOnSuccessListener { authResult ->
-                Log.d(TAG, "firebaseAuthWithGoogleAccount : logged in")
-
-                val firebaseUser = firebaseAuth.currentUser
-                val uid = firebaseUser!!.uid
-                val email = firebaseUser.email
-                startActivity(Intent(this, MapActivity::class.java))
-            }
-    }
-
-    /*
-    fun Maps(view: View?){
-        val intent = Intent(this, MapsActivity::class.java)
-
-        startActivity(intent)
-    }*/
-
-
-    fun goToMap(view: View?){
-        val intent = Intent(this, MapActivity::class.java)
-        startActivity(intent)
-    }
-}
