+++ app/src/main/java/supportClasses/TabPageAdapter.kt
@@ -2,10 +2,10 @@ package supportClasses
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import mx.itesm.bamx.activities.RootTabActivity
-import mx.itesm.bamx.activities.fragments.DonationFragment
-import mx.itesm.bamx.activities.fragments.MapFragment
-import mx.itesm.bamx.activities.fragments.RewardsFragment
+import mx.itesm.bamx.RootTabActivity
+import mx.itesm.bamx.fragments.DonationFragment
+import mx.itesm.bamx.fragments.MapFragment
+import mx.itesm.bamx.fragments.RewardsFragment
 
 
 class TabPageAdapter(activity: RootTabActivity, private val tabCount : Int) : FragmentStateAdapter(activity) {
